backend/requirements.txt
africastalking==1.2.9
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-bootstrap4==24.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-js-asset==3.0.1
django-tenants==3.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
schema==0.7.7
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
whitenoise==6.9.0
